Split out ReactART setup for Reconciler injection
